Move job scripts & update refs
