+++ dev.dockerfile
@@ -33,6 +33,9 @@ COPY run_command.sh run_command.sh
 COPY run.py run.py
 COPY system_profile_validator.py system_profile_validator.py
 COPY generate_stale_host_notifications.py generate_stale_host_notifications.py
+COPY create_ungrouped_host_groups.py create_ungrouped_host_groups.py
+COPY export_group_data_s3.py export_group_data_s3.py
+COPY update_hosts_last_check_in.py update_hosts_last_check_in.py
 RUN chown -R 1001:0 ./
 USER 1001
